BRD-MergedTenantMaster-Backend/brd_platform/requirements.txt
Django==4.2.7
djangorestframework==3.20.0
djangorestframework-simplejwt==5.2.0
django-cors-headers==4.0.0
django-filter==24.0
gunicorn==21.2.0
